refactor:  zod 스키마를 타입으로 선언하고 코드를 분리하였습니다.
