--- src/types/receiver.ts
@@ -1,5 +1,0 @@
-export type Receiver = {
-  receiverName: string;
-  receiverPhoneNumber: string;
-  itemCount: number;
-};
